requirements.txt
python-dotenv
requests
selenium
m3u8